update browser & js protocol to 1.3

+++ src/main/java/io/webfolder/cdp/event/emulation/VirtualTimePaused.java
@@ -28,19 +28,21 @@ import io.webfolder.cdp.annotation.Experimental;
 @Domain("Emulation")
 @EventName("virtualTimePaused")
 public class VirtualTimePaused {
-    private Integer virtualTimeElapsed;
+    private Double virtualTimeElapsed;
 
     /**
-     * The amount of virtual time that has elapsed in milliseconds since virtual time was first enabled.
+     * The amount of virtual time that has elapsed in milliseconds since virtual time was first
+     * enabled.
      */
-    public Integer getVirtualTimeElapsed() {
+    public Double getVirtualTimeElapsed() {
         return virtualTimeElapsed;
     }
 
     /**
-     * The amount of virtual time that has elapsed in milliseconds since virtual time was first enabled.
+     * The amount of virtual time that has elapsed in milliseconds since virtual time was first
+     * enabled.
      */
-    public void setVirtualTimeElapsed(Integer virtualTimeElapsed) {
+    public void setVirtualTimeElapsed(Double virtualTimeElapsed) {
         this.virtualTimeElapsed = virtualTimeElapsed;
     }
 }
